+++ CHANGELOG.md
@@ -6,6 +6,7 @@ PowerModels.jl Change Log
 - Fix dictionary call syntax error in psse export (#941)
 - Updated the logging message when components are deactivated (#943)
 - Improve performance of `calc_basic_incidence_matrix` (#946)
+- Fix logic mistake in resolve switches and add test (#956)
 - Fix `parse_file` to use `JSON.parsefile` (#958)
 - Fix spelling of `resolve_swithces!` and add deprecation for compatibility (#959)